rm coordinates_views
