+++ mmda-backend/backend/views/__init__.py
@@ -6,7 +6,6 @@ Import the blueprints here
 from .admin_views import admin_blueprint
 from .collocation_views import collocation_blueprint
 from .constellation_views import constellation_blueprint
-from .coordinates_views import coordinates_blueprint
 from .corpus_views import corpus_blueprint
 from .discourseme_views import discourseme_blueprint
 from .keyword_views import keyword_blueprint
@@ -26,5 +25,4 @@ def register_blueprints(app):
     app.register_blueprint(collocation_blueprint)
     app.register_blueprint(keyword_blueprint)
     app.register_blueprint(discourseme_blueprint)
-    app.register_blueprint(coordinates_blueprint)
     app.register_blueprint(constellation_blueprint)
